refactor(app-web): move workspace Gemini BYO key into the Models section (hosted)
Extract the ws-llm-key body into sections/llm-key-block.tsx; hosted embeds
it in Settings -> Models (one home for model-related settings), OSS keeps
the standalone ws-llm-key section since it has no Models entry.
